Address review comments
- Move rule download logic to launch.rulesengine
- Rename RulesDownloader to RulesLoader
- Add loadFromUrl , loadFromCache and loadFromAsset to RulesLoader
- Rename RulesDownloadResult to RulesLoadResult
- Change logs to trace when logging failure
- Add more tests

+++ code/android-core-library/src/main/java/com/adobe/marketing/mobile/launch/rulesengine/download/RulesZipProcessingHelper.java
@@ -1,4 +1,4 @@
-package com.adobe.marketing.mobile.rulesengine.download;
+package com.adobe.marketing.mobile.launch.rulesengine.download;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
@@ -41,11 +41,11 @@ class RulesZipProcessingHelper {
      * @return true if the directory creation is successful; false otherwise.
      */
     boolean createTemporaryRulesDirectory(final String tag) {
-        if (StringUtils.isNullOrEmpty(tag)) return false;
+        if (!canProcess(tag)) return false;
 
         final File tempExtractDir = getTemporaryDirectory(tag);
         if (!tempExtractDir.exists() && !tempExtractDir.mkdirs()) {
-            Log.trace(CoreConstants.LOG_TAG, TAG, "Cannot access application cache directory to create temp dir.");
+            Log.debug(CoreConstants.LOG_TAG, TAG, "Cannot access application cache directory to create temp dir.");
             return false;
         }
         return true;
@@ -60,12 +60,12 @@ class RulesZipProcessingHelper {
      * @return true if the zip stream was successfully copied; false otherwise.
      */
     boolean storeRulesInTemporaryDirectory(final String tag, final InputStream zip) {
-        if (StringUtils.isNullOrEmpty(tag)) return false;
+        if (!canProcess(tag)) return false;
 
         final boolean fileRead = FileUtils.readInputStreamIntoFile(getZipFileHandle(tag), zip, false);
 
         if (!fileRead) {
-            Log.trace(CoreConstants.LOG_TAG, TAG, "Cannot read response content into temp dir.");
+            Log.debug(CoreConstants.LOG_TAG, TAG, "Cannot read response content into temp dir.");
             return false;
         }
 
@@ -79,14 +79,14 @@ class RulesZipProcessingHelper {
      * @return the contents of TEMP_RULES_JSON if it was found and was readable; null otherwise.
      */
     String unzipRules(final String tag) {
-        if (StringUtils.isNullOrEmpty(tag)) return null;
+        if (!canProcess(tag)) return null;
 
         final File temporaryDirectory = getTemporaryDirectory(tag);
         final boolean extracted =
                 FileUtils.extractFromZip(getZipFileHandle(tag), temporaryDirectory.getPath());
 
         if (!extracted) {
-            Log.trace(CoreConstants.LOG_TAG, TAG, "Failed to extract rules response zip into temp dir.");
+            Log.debug(CoreConstants.LOG_TAG, TAG, "Failed to extract rules response zip into temp dir.");
             return null;
         }
 
@@ -94,7 +94,7 @@ class RulesZipProcessingHelper {
                 new File(temporaryDirectory.getPath() + File.separator + TEMP_RULES_JSON);
 
         if (!extractedRulesJsonFile.exists()) {
-            Log.trace(CoreConstants.LOG_TAG, TAG, "Extract rules directory does not contain a rules.json file.");
+            Log.debug(CoreConstants.LOG_TAG, TAG, "Extract rules directory does not contain a rules.json file.");
             return null;
         }
 
@@ -102,13 +102,13 @@ class RulesZipProcessingHelper {
             final String rules = StreamUtils.readAsString(fileInputStream);
 
             if (rules == null) {
-                Log.trace(CoreConstants.LOG_TAG, TAG, "Null content from rules.json file.");
+                Log.debug(CoreConstants.LOG_TAG, TAG, "Null content from rules.json file.");
                 return null;
             }
 
             return rules;
         } catch (final IOException e) {
-            Log.trace(CoreConstants.LOG_TAG, TAG,
+            Log.debug(CoreConstants.LOG_TAG, TAG,
                     "Exception while processing rules from source %s", tag);
             return null;
         }
@@ -119,18 +119,24 @@ class RulesZipProcessingHelper {
      * @param tag the tag used to create that
      */
     void deleteTemporaryDirectory(final String tag) {
+        if (!canProcess(tag)) return;
+
         FileUtils.deleteFile(getTemporaryDirectory(tag), true);
     }
 
     @VisibleForTesting
-    File getTemporaryDirectory(@NonNull final String key) {
-        final String hash = StringEncoder.sha2hash(key);
+    File getTemporaryDirectory(@NonNull final String tag) {
+        final String hash = StringEncoder.sha2hash(tag);
         return new File(deviceInfoService.getApplicationCacheDir().getPath()
                 + File.separator + TEMP_DOWNLOAD_DIR
                 + File.separator + hash);
     }
 
-    private File getZipFileHandle(@NonNull final String key) {
-        return new File(getTemporaryDirectory(key).getPath() + File.separator + TEMP_RULES_ZIP);
+    private File getZipFileHandle(@NonNull final String tag) {
+        return new File(getTemporaryDirectory(tag).getPath() + File.separator + TEMP_RULES_ZIP);
+    }
+
+    private boolean canProcess(@NonNull final String tag) {
+        return !StringUtils.isNullOrEmpty(tag);
     }
 }
